Make iamc.data helper functions available in DB layer
